+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.32.0
 * Add `packageTracking` method to `TransactionGateway` to make request to add tracking information to transactions
 * Add `PackageDetails` and `PackageTrackingRequest` classes
 * Add `process_debit_as_credit` to `TransactionOptionsCreditCardRequest`